requirements.txt
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.4.0
transformers==4.37.2
torch==2.1.2
fastapi==0.108.0
uvicorn==0.25.0
websockets==12.0
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.3
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
nltk==3.8.1
sentence-transformers==2.2.2
scikit-learn==1.4.0
jupyter==1.0.0
ipykernel==6.29.0
reportlab==4.0.9
pdfkit==1.0.0
markdown==3.5.2
asyncio==3.4.3
aiohttp==3.9.1
beautifulsoup4==4.12.3
